+++ tests/unit/models/test_task.py
@@ -1,6 +1,7 @@
 from pathlib import Path
 
 import pytest
+from pydantic import ValidationError
 
 from dbx.models.cli.execute import ExecuteParametersPayload
 from dbx.models.workflow.common.task import SparkPythonTask, SparkJarTask, SparkSubmitTask, BaseTaskMixin
@@ -34,6 +35,13 @@ def test_spark_python_task_positive(temp_project: Path):
     assert isinstance(_t.execute_file, Path)
 
 
+def test_spark_python_task_non_py_file(temp_project: Path):
+    py_file = f"file://{temp_project.name}/tasks/sample_etl_task.ipynb"
+    _payload = get_spark_python_task_payload(py_file).get("spark_python_task")
+    with pytest.raises(ValidationError):
+        SparkPythonTask(**_payload)
+
+
 def test_sql_task_non_unique():
     payload = {"query": {"query_id": "some"}, "dashboard": {"dashboard_id": "some"}, "warehouse_id": "some"}
     with pytest.raises(ValueError):
@@ -63,7 +71,7 @@ def test_spark_python_task_not_fuse(temp_project: Path):
 def test_spark_python_task_execute_incorrect(temp_project: Path):
     py_file = f"file://{temp_project.name}/tasks/sample_etl_task.py"
     _payload = get_spark_python_task_payload(py_file).get("spark_python_task")
-    _payload["python_file"] = "dbfs:/some/path"
+    _payload["python_file"] = "dbfs:/some/path.py"
     with pytest.raises(ValueError):
         _st = SparkPythonTask(**_payload)
         _st.execute_file  # noqa